Updated Layout, New login page, Message Bar separated
Basically I gave the input box its own div but have yet to get the chat section div to scroll down when a new message is saved, I added a simple login page that's just a popup and tells the user to sign in with google, I changed the color scheme to match somewhat the colors of Kean.

What I'm adding next:
Input box expanded to fit the div
Logo on the navigation bar
Differentiation between sent and received.